Create README.md
The Rainwatch Africa portal is developing projects for monitoring rainfall and its distribution for the African Continent. Currently just a handful of countries are 
involved. The portal requires entry data to conform to a sppecific format. The project was to wrangle that of the Zimbabwe Meteorological Agency for readiness for the Rain-watch Africa
portal